+++ packages/apple_vision_recognize_text/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+
+* Added language support
+* Added automatic language detection
+* Added recognition level
+
 ## 0.0.2
 
 * Added image orientation